Move RFC example tests into compliance suite
