+++ tests/compliance.rs
@@ -1,0 +1,4 @@
+#[path = "compliance/commands.rs"]
+mod commands;
+#[path = "compliance/rfc_e2e.rs"]
+mod rfc_e2e;
