+++ tests/integration.rs
@@ -1,0 +1,18 @@
+#[path = "integration/auth.rs"]
+mod auth;
+#[path = "integration/cancel_lock.rs"]
+mod cancel_lock;
+#[path = "integration/control.rs"]
+mod control;
+#[path = "integration/moderated.rs"]
+mod moderated;
+#[path = "integration/peers.rs"]
+mod peers;
+#[path = "integration/retention.rs"]
+mod retention;
+#[path = "integration/max_size.rs"]
+mod max_size;
+#[path = "integration/storage.rs"]
+mod storage;
+#[path = "integration/tls.rs"]
+mod tls;
