+++ tests/integration/auth.rs
