+++ tests/integration/cancel_lock.rs
