+++ tests/unit.rs
@@ -1,0 +1,4 @@
+#[path = "unit/config.rs"]
+mod config;
+#[path = "unit/wildmat.rs"]
+mod wildmat;
